Visually center card(s) on X-Axis, fix overflow
Previously, due to the chosen values for `margin`, the card-container
was overflowing and not visually centered. Card Contents also were not
centered due to a 0.5em margin on the left. This was especially
noticeable on the new page that has images with a border, making it very
evident that the contents were offcenter.

Though, these decisions might have been made with a different intent for
the design, as such these (admittely small) changes may not fit the
desired look.